actions: Uses a different account for backports and autolabel backports
The github-actions bot does not trigger any other CI workflows, but we'd
like to have tests running for the backports as well. Which is why the
backports will be made with a different account.

Additionally, we're now adding the "automerge" label to the new
backports.

(cherry picked from commit 813a29b8a5a5ab4178481adc2a8c46995ae4bd22)